refactor(project): update project structure
