--- src/html_tracing/utilities/__init__.py
@@ -1,1 +1,0 @@
-"""Package Utilities."""
