+++ src/utilities/__init__.py
@@ -1,0 +1,8 @@
+"""Package Utilities."""
+
+
+from .logger import Logger
+
+__all__ = ["Logger"]
+
+logger = Logger(tracing=True)
